requirements.txt
numpy
pandas
matplotlib
networkx
fastapi[standard]>=0.112.0
uvicorn
jinja2
plotly
psycopg2-binary